+++ frontend/src/app/modules/shared/assembler/parts.assembler.spec.ts
@@ -106,11 +106,12 @@ describe('PartsAssembler', () => {
     const productionDate = 'productionDate';
     const qualityType = 'qualityType';
     const serialNumber = 'serialNumber';
+    const semanticDataModel = 'semanticDataModel';
 
     it('should clean up data for part view', () => {
-      const data = { productionDate, qualityType, serialNumber, test: '' } as unknown as Part;
+      const data = { productionDate, qualityType, serialNumber, semanticDataModel, test: '' } as unknown as Part;
       expect(PartsAssembler.filterPartForView({ data })).toEqual({
-        data: { name: undefined, productionDate, serialNumber } as unknown as Part,
+        data: { name: undefined, productionDate, serialNumber, semanticDataModel } as unknown as Part,
       });
     });